% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcsmc_bulk_samples.R
\name{ABCSMCwithBulkSamples}
\alias{ABCSMCwithBulkSamples}
\title{ABC SMC with bulk tumour samples}
\usage{
ABCSMCwithBulkSamples(sim.id = 1, grid = 50, mu.rl = 10, s.rl = 2,
  t.rl = 3, d.rl = 0, a.rl = 1, random.state = 123, depth = 100,
  min.reads = 5, depth.model = 1, nbulks = 3, radius = 5,
  nbins = 10, dimension = "2D", output.dir = "data/inferred_params/",
  sampling = "needle", params.to.rec = "mu_t_s_d_a", sigma = 0.5,
  N = 100, rounds = 10, ncores = 2, params = data.table(name =
  c("mu", "t", "s", "d", "a"), min = c(1, 1, 1, 0, 0), max = c(100, 20, 5,
  1, 1), fix = c(F, F, F, F, F)), target.data = NULL,
  target.popsize = NULL)
}
\arguments{
\item{sim.id}{Simulated tumour ID.}

\item{grid}{Tumour space grid size.}

\item{mu.rl}{Mutation rate.}

\item{s.rl}{Selective advantage.}

\item{t.rl}{Time to a new subclone introduction.}

\item{d.rl}{Death rate.}

\item{a.rl}{Aggression (cell pushing probability scaled by its location on the grid).}

\item{random.state}{Random state.}

\item{depth}{Sequencing depth.}

\item{min.reads}{Minimum number of detectable reads.}

\item{depth.model}{Model for the sequencing depth. 1: Poisson distributed depth (default). 2: Beta-binomial distributed depth. 3: Fixed depth.}

\item{nbulks}{Number of bulks to sample.}

\item{radius}{Bulk radius.}

\item{nbins}{Number of bins for a multivariate histogram construction.}

\item{dimension}{Tumour space grid dimension.}

\item{output.dir}{Path to a directory where the output files will be exported.}

\item{sampling}{Sampling scheme.}

\item{params.to.rec}{String of parameter names (will be set as a directory name where the inferred particle files will be saved per SMC round).}

\item{sigma}{Permutation kernel scale factor.}

\item{N}{Number of simulations to run per SMC round.}

\item{rounds}{Number of SMC rounds.}

\item{ncores}{Number of cores to run in parallel with.}

\item{params}{Data.table containing parameter names, uniform prior distribution limits and flags to fix each parameter at given (previously inferred or already known) values or not. For exact format of the table, please see the examples section.}

\item{target.data}{List of bulk samples (each element should be named as "mutation_data"; so an access call to a bulk sample element would be: bulk.data.list[[id]]$mutation_data). Each bulk sample should be formatted as a data.frame of unique mutations with the following columns: clone (Clone type label), alt (Number of reads), depth (Sequencing depth), id (Unique ID).}

\item{target.popsize}{Estimated size (as number of cells) of the tumour to infer the parameters for.}
}
\value{
Exports inferred set of parameters by each SMC round (as .csv files) into a given directory.
}
\description{
Performs ABC SMC simulations to recover a given parameter set
   for a tumour simulation using its bulk samples.
}
\author{
Kate Chkhaidze, \email{kate.chkhaidze@icr.ac.uk}
}
