% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_utilities.R
\name{CheckRootBinary}
\alias{CheckRootBinary}
\title{Make unrooted non-binary phylogenetic tree rooted and binary}
\usage{
CheckRootBinary(tree)
}
\arguments{
\item{tree}{ape phylo object}
}
\value{
Rooted binary version of a given phylogenetic tree.
}
\description{
Checks if a given phylo object is rooted and binary. If not, converts and returns an updated tree.
}
\author{
Kate Chkhaidze, \email{kate.chkhaidze@icr.ac.uk}
}
