% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{GetCumSums}
\alias{GetCumSums}
\title{VAF cumulative sums}
\usage{
GetCumSums(vaf, fmin = 0.05, fmax = 0.7, step.size = 0.01)
}
\arguments{
\item{vaf}{Variant Allele Frequency of a sample.}

\item{fmin}{Minimum frequency to include.}

\item{fmax}{Maximum frequency to include.}

\item{step.size}{Step size for a cumulative walk.}
}
\value{
Vector of cumulative sums.
}
\description{
Calculates cumulative sums for a given VAF array.
}
\author{
Kate Chkhaidze, \email{kate.chkhaidze@icr.ac.uk}
}
