% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{GetDensityCounts}
\alias{GetDensityCounts}
\title{VAF density counts}
\usage{
GetDensityCounts(vaf, nbreaks = seq(0, 1, 0.01))
}
\arguments{
\item{vaf}{Variant Allele Frequency of a sample.}

\item{nbreaks}{Number of bins.}
}
\value{
Vector of histogram counts.
}
\description{
Extracts bin counts from a given VAF histogram.
}
\author{
Kate Chkhaidze, \email{kate.chkhaidze@icr.ac.uk}
}
