% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_utilities.R
\name{GetMode}
\alias{GetMode}
\title{Find mode of a numeric or character vector}
\usage{
GetMode(x)
}
\arguments{
\item{x}{Numeric or character vector.}
}
\value{
Mode of x.
}
\description{
Given a vector of numeric or character values, returns its mode (most frequent values). If all elements have a same frequency, returns NA.
}
\author{
Kate Chkhaidze, \email{kate.chkhaidze@icr.ac.uk}
}
