% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{GetMultivrtDistr}
\alias{GetMultivrtDistr}
\title{Multivariate distribution of multiple bulk samples}
\usage{
GetMultivrtDistr(bulk.data.list, nbins, nbulks, add.clonal.muts = F)
}
\arguments{
\item{bulk.data.list}{List of sampled bulk mutations.}

\item{nbins}{Number of bins to build a multivariate histogram with.}

\item{nbulks}{Number of bulk samples.}

\item{add.clonal.muts}{Boolean to add random clonal mutations or not.}
}
\value{
Multivariate distribution of VAFs obtained for a given list of bulk samples.
}
\description{
Calculates multivariate joint distribution for a set of sampled
   bulk mutations.
}
\author{
Kate Chkhaidze, \email{kate.chkhaidze@icr.ac.uk}
}
