% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_utilities.R
\name{PerturbUniKernel}
\alias{PerturbUniKernel}
\title{Uniform perturbation kernel}
\usage{
PerturbUniKernel(value, maxval, minval, frac = 2)
}
\arguments{
\item{value}{Parameter value to perturb.}

\item{maxval}{Upper limit for the uniform distribution.}

\item{minval}{Lower limit for the uniform distribution.}

\item{frac}{Scale factor for sigma.}
}
\value{
Perturbed parameter value.
}
\description{
Perturbs a given parameter value with a corresponding
   uniform kernel.
}
\author{
Kate Chkhaidze, \email{kate.chkhaidze@icr.ac.uk}
}
