% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_schemes.R
\name{PrepareBiopsyNeedles}
\alias{PrepareBiopsyNeedles}
\title{Get needle biopsy coordinates}
\usage{
PrepareBiopsyNeedles(radius, grid, nsamples, dimension)
}
\arguments{
\item{radius}{Needle width radius.}

\item{grid}{Tumour space grid size.}

\item{nsamples}{Number of biopsies to sample.}

\item{dimension}{Tumour space grid dimension.}
}
\value{
List of cell coordinates for sampling.
}
\description{
Prepares coordinates for cells inside a given needle shaped
     biopsies to be then sampled from a simulated tumour.
}
\author{
Kate Chkhaidze, \email{kate.chkhaidze@icr.ac.uk}
}
