% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_schemes.R
\name{PrepareBiopsySquares}
\alias{PrepareBiopsySquares}
\title{Get punch biopsy coordinates}
\usage{
PrepareBiopsySquares(radius, grid, dimension)
}
\arguments{
\item{radius}{Square side times 1/2.}

\item{grid}{Tumour space grid size.}

\item{dimension}{Tumour space grid dimension.}
}
\value{
List of cell coordinates for sampling.
}
\description{
Prepares coordinates for cells inside a given square shaped
     biopsies to be then sampled from a simulated tumour.
}
\author{
Kate Chkhaidze, \email{kate.chkhaidze@icr.ac.uk}
}
