% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_schemes.R
\name{PrepareSingleCellSamples}
\alias{PrepareSingleCellSamples}
\title{Get single cell sample coordinates}
\usage{
PrepareSingleCellSamples(x, y, z, nsamples, dimension)
}
\arguments{
\item{x}{Tumour space grid dimension 1.}

\item{y}{Tumour space grid dimension 2.}

\item{z}{Tumour space grid dimension 3.}

\item{nsamples}{Number of single cells to sample.}

\item{dimension}{Tumour space grid dimension.}
}
\value{
List of single cell coordinates to sample.
}
\description{
Prepares coordinates for single cells to be sampled
   from a simulated tumour.
}
\author{
Kate Chkhaidze, \email{kate.chkhaidze@icr.ac.uk}
}
