% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SimulateTumourSample}
\alias{SimulateTumourSample}
\title{Simulate tumour bulk samples}
\usage{
SimulateTumourSample(x = 100, y = 100, z = 1, birthrates = 1,
  deathrates = 0, aggressions = 0, push_power = max(x, y, z),
  mutation_rates = 10, clone_start_times = 0, kill_regrow_times = 1,
  father = 0, clonal_mutations = 0, seed = 42, verbose = FALSE,
  depth = 100, min_reads = 2, min_vaf = 0, depth_model = 1,
  sample_strategy = list(c(1, 1, 1, x, y, z)))
}
\arguments{
\item{x}{Space size in dimension one.}

\item{y}{Space size in dimension two. If missing model is 1D.}

\item{z}{Space size in dimension three. If missing model is 2D.}

\item{birthrates}{Birth rates.}

\item{deathrates}{Death rates.}

\item{aggressions}{Aggression probability allowing cells to push or not.}

\item{push_power}{Cells' pushing probability scaled by their location on the grid.}

\item{mutation_rates}{Mutation rates.}

\item{clone_start_times}{Times to new subclone introduction.}

\item{kill_regrow_times}{Times to kill 99 percent of tumour cell population and regrow the tumour.}

\item{father}{Father cell ID.}

\item{clonal_mutations}{Number of clonal mutations to add post simulation and sampling.}

\item{seed}{Random state.}

\item{verbose}{Prints simulation progress.}

\item{depth}{Sequencing depth.}

\item{min_reads}{Minimum number of detectable reads.}

\item{min_vaf}{Filters mutations by given minimum VAF.}

\item{depth_model}{Depth model - fixed vs sampling sequencing depth from a Poisson distribuition with mean given by the 'depth' argument.}

\item{sample_strategy}{Bulk sample coordinates.}
}
\value{
List of matrices (corresponding to each bulk) of sampled and sequenced mutations.
}
\description{
Simulates tumour growth followed by bulk sampling and exporting sequenced list of mutations.
}
\author{
Timon Heide, \email{timon.heide@icr.ac.uk}, Kate Chkhaidze, \email{kate.chkhaidze@icr.ac.uk}
}
