% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{SimulateTumourTree}
\alias{SimulateTumourTree}
\title{Simulate tumour single cell sampling phylogenetic trees}
\usage{
SimulateTumourTree(x = 100, y = 100, z = 1, birthrates = 1,
  deathrates = 0, aggressions = 1, push_power = max(x, y, z),
  mutation_rates = 10, clone_start_times = 0, kill_regrow_times = 1,
  father = 0, clonal_mutations = 0, seed = 42, verbose = FALSE,
  strategy = NULL, nsamples = 10)
}
\arguments{
\item{x}{Space size in dimension one.}

\item{y}{Space size in dimension two. If missing model is 1D.}

\item{z}{Space size in dimension three. If missing model is 2D.}

\item{birthrates}{Birth rates.}

\item{deathrates}{Death rates.}

\item{aggressions}{Aggression probability allowing cells to push or not.}

\item{push_power}{Cells' pushing probability scaled by their location on the grid.}

\item{mutation_rates}{Mutation rates.}

\item{clone_start_times}{Times to new subclone introduction.}

\item{kill_regrow_times}{Times to kill 99 percent of tumour cell population and regrow the tumour.}

\item{father}{Father cell ID.}

\item{clonal_mutations}{Number of clonal mutations to add post simulation and sampling.}

\item{seed}{Random state.}

\item{verbose}{Prints simulation progress.}

\item{strategy}{Single-cell sample coordinates.}

\item{nsamples}{Number of single-cell samples.}
}
\value{
List of matrices (corresponding to each bulk) of sampled and sequenced mutations.
}
\description{
Simulates tumour growth followed by single cell sampling and constructing corresponding phylogenetyic trees.
}
\author{
Timon Heide, \email{timon.heide@icr.ac.uk}, Kate Chkhaidze, \email{kate.chkhaidze@icr.ac.uk}
}
